requirements.txt
typing
crewai
fastapi
pydantic
pathlib
evolutionapi
uvicorn
asyncpg
supabase
mysql-connector-python
pandas
psycopg2